requirements.txt
Flask>=2.0.1
Flask-PyMongo>=2.3.0
flask-simplelogin>=0.1.0
flask-cors>=3.0.10
docker>=5.0.0
pymongo>=3.12.0
flask_simplelogin>=0.1.0
paho-mqtt>=1.5.1